requirements.txt
pysimplegui~=4.56.0
pillow~=9.0.1